Add filtering flag to CNVs
CNVs are grouped across callers according to overlap. If at least one
CNV in the group is part of the set of filtered CNVs, the whole group
passes the filter. The reason for this is to be able to display CNVs
that are not picked from a specific caller, but is called by another
caller.