fix: delegate schema validation to reports module
Due to if-then not being a part of Draft-4 of JSON Schema, I had to
update also this schema to Draft-7 since this apparently gets applied to
module schemas as well. As far as I can tell, this should be fully
backwards compatible, at least the features being used in this schema.